test: separate unit & smoke tests
